meta.in: Compact all the code cells before evaluating.
This compacts all the cells by 31 (the biggest special ASCII char not
used on modern systems). Because of that, one saves:
- CPU cycles (less looped operations on code cells);
- Code chars (meta.r.min.bf shortens from 1864 to 1376 bytes---crazy!);
- And all that without sacrificing readability (dbfi.b compacts cells
  too, but uses a much less readable, understandable, and maintainable
  array of differences approach).

I mean, meta.bf is still not the level of dbfi compactness (423
bytes). But maybe it's possible for meta.bf to reach at least 1000
bytes, still fulfilling all the requirements set for it?